Added add to cart feature on the product details page
