+++ src/App.jsx
@@ -1,3 +1,5 @@
+import { useState } from "react";
+
 import { PageNotFound } from "components/commons";
 import { Route, Switch, Redirect } from "react-router-dom";
 import routes from "routes";
@@ -5,13 +7,20 @@ import routes from "routes";
 import "./App.css";
 import Product from "./components/Product";
 import ProductList from "./components/ProductList";
+import CartItemsContext from "./contexts/CartItemsContext";
 
-const App = () => (
-  <Switch>
-    <Route exact component={Product} path={routes.products.show} />
-    <Route exact component={ProductList} path={routes.products.index} />
-    <Redirect exact from={routes.root} to={routes.products.index} />
-    <Route component={PageNotFound} path="*" />
-  </Switch>
-);
+const App = () => {
+  const [cartItems, setCartItems] = useState([]);
+
+  return (
+    <CartItemsContext.Provider value={[cartItems, setCartItems]}>
+      <Switch>
+        <Route exact component={Product} path={routes.products.show} />
+        <Route exact component={ProductList} path={routes.products.index} />
+        <Redirect exact from={routes.root} to={routes.products.index} />
+        <Route component={PageNotFound} path="*" />
+      </Switch>
+    </CartItemsContext.Provider>
+  );
+};
 export default App;
