+++ src/components/Product/index.jsx
@@ -1,6 +1,7 @@
 import { useState, useEffect } from "react";
 
 import productsApi from "apis/products";
+import AddToCart from "components/AddToCart";
 import { Header, PageLoader, PageNotFound } from "components/commons";
 import { Typography } from "neetoui";
 import { append, isNotNil } from "ramda";
@@ -65,6 +66,7 @@ const Product = () => {
           <Typography className="font-semibold text-green-600">
             {discountPercentage}% off
           </Typography>
+          <AddToCart {...{ slug }} />
         </div>
       </div>
     </>
