+++ src/components/ProductList/ProductListItem.jsx
@@ -4,14 +4,7 @@ import { Link } from "react-router-dom";
 import routes from "routes";
 import { buildUrl } from "utils/url";
 
-const ProductListItem = ({
-  imageUrl,
-  name,
-  offerPrice,
-  slug,
-  isInCart,
-  toggleIsInCart,
-}) => (
+const ProductListItem = ({ imageUrl, name, offerPrice, slug }) => (
   <Link
     className="neeto-ui-border-black neeto-ui-rounded-xl flex w-48 flex-col items-center justify-between border p-4"
     to={buildUrl(routes.products.show, { slug })}
@@ -21,7 +14,7 @@ const ProductListItem = ({
       {name}
     </Typography>
     <Typography>${offerPrice}</Typography>
-    <AddToCart {...{ isInCart, toggleIsInCart }} />
+    <AddToCart {...{ slug }} />
   </Link>
 );
