+++ src/contexts/CartItemsContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const CartItemsContext = createContext();
+
+export default CartItemsContext;
